requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator
